+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * Added a new flag to **generate-docs** command, allowing to add a custom image link to a playbook README.
 * Added a new validation to the **validate** command to verify that the package directory name is the same as the files contained in the that package.
 * Added support in the **unify** command to unify a schema into its Modeling Rule.
+* The **format** command will now replace the *fromServerVersion* field with *fromVersion*.
 
 ## 1.6.8